tests(metrics): refactor FC metrics validation functions
To have clear distintion between CW specific metrics functions and
Firecracker specific metrics functions, move fcmetrics to its own file.
